+++ flowclient/flowclient/errors.py
@@ -1,0 +1,4 @@
+class FlowclientConnectionError(Exception):
+    """
+    Custom exception to indicate an error when connecting to a FlowKit API.
+    """
